separate device util files
