+++ src/otx/utils/utils.py
@@ -13,17 +13,13 @@ from functools import partial
 from types import LambdaType
 from typing import TYPE_CHECKING, Any, Callable
 
-import torch
-
 from otx.core.model.base import OTXModel
 
 if TYPE_CHECKING:
     from pathlib import Path
 
+    import torch
     from jsonargparse import Namespace
-
-
-XPU_AVAILABLE = None
 
 
 def get_using_dot_delimited_key(key: str, target: Any) -> Any:  # noqa: ANN401
@@ -128,14 +124,6 @@ def remove_matched_files(directory: Path, pattern: str, file_to_leave: Path | No
     for weight in directory.rglob(pattern):
         if weight != file_to_leave:
             weight.unlink()
-
-
-def is_xpu_available() -> bool:
-    """Checks if XPU device is available."""
-    global XPU_AVAILABLE  # noqa: PLW0603
-    if XPU_AVAILABLE is None:
-        XPU_AVAILABLE = hasattr(torch, "xpu") and torch.xpu.is_available()
-    return XPU_AVAILABLE
 
 
 def get_model_cls_from_config(model_config: Namespace) -> type[OTXModel]:
